Change DOC link
Previously we were linking contributors directly to the DCO text on the
web but honestly that doesn't do a good job of explainging what a DCO is
or the process. I've updated the link to go to DCO bot, which explains
both what it means to sign, how to do it and how the bot works.

Signed-off-by: Carolyn Van Slyck <me@carolynvanslyck.com>